PCE: Fixed Scaling broken since 1.26a2

+++ huexpress-go/main/main.c
@@ -93,8 +93,8 @@ void osd_gfx_init(void)
 void osd_gfx_set_mode(int width, int height)
 {
     const rg_display_t *display = rg_display_get_status();
-    int crop_h = MAX(0, width - display->screen.width);
-    int crop_v = MAX(0, height - display->screen.height) + (overscan ? 6 : 0);
+    int crop_h = MAX(0, width - (int)display->screen.width);
+    int crop_v = MAX(0, height - (int)display->screen.height) + (overscan ? 6 : 0);
 
     // We center the content vertically and horizontally to allow overflows all around
     int offset_center = (((XBUF_HEIGHT - height) / 2 + 16) * XBUF_WIDTH + (XBUF_WIDTH - width) / 2);
